+++ controllers/PersonController.php
@@ -7,33 +7,45 @@ use yii\web\Controller;
 use app\models\CoachModel;
 use app\models\Person;
 
-class PersonController extends BaseController {
+class PersonController extends BaseController
+{
 
     public $layout = 'inner';
 
-    public function actionIndex() {
+    public function actionIndex()
+    {
         $persons = Person::browse();
         return $this->render('index', [
                     'persons' => $persons,
         ]);
     }
 
-    public function actionView($id) {
+    public function actionView($id)
+    {
         $person = Person::findOne(['id' => $id]);
+
+        if (!$person || $person->coach_id != Yii::$app->user->id) {
+            throw new \yii\web\ForbiddenHttpException(Yii::t('app', 'Your not allowed to access this page.'));
+        }
         Yii::$app->session->set('person_id', $id);
         return $this->render('view', [
                     'person' => $person,
         ]);
     }
 
-    public function actionNew() {
+    public function actionNew()
+    {
         $person = new Person();
 
-        if ($person->load(Yii::$app->request->post()) && $person->save()) {
-            SiteController::addFlash('success', Yii::t('app', '{name} has been successfully created.', ['name' => $person->fullname]));
-            return $this->redirect(['/person']);
-        } else {
-            SiteController::FlashErrors($person);
+        if ($person->load(Yii::$app->request->post())) {
+            $this->upload($person, 'photo');
+
+            if ($person->save()) {
+                SiteController::addFlash('success', Yii::t('app', '{name} has been successfully created.', ['name' => $person->fullname]));
+                return $this->redirect(['/person']);
+            } else {
+                SiteController::FlashErrors($person);
+            }
         }
 
         return $this->render('form', [
@@ -41,14 +53,23 @@ class PersonController extends BaseController {
         ]);
     }
 
-    public function actionEdit($id) {
+    public function actionEdit($id)
+    {
         $person = Person::findOne(['id' => $id]);
 
-        if ($person->load(Yii::$app->request->post()) && $person->save()) {
-            SiteController::addFlash('success', Yii::t('app', '{name} has been successfully edited.', ['name' => $person->fullname]));
-            return $this->redirect(['/person']);
-        } else {
-           SiteController::FlashErrors($person);
+        if (!$person || $person->coach_id != Yii::$app->user->id) {
+            throw new \yii\web\ForbiddenHttpException(Yii::t('app', 'Your not allowed to access this page.'));
+        }
+
+        if ($person->load(Yii::$app->request->post())) {
+            self::upload($person, 'photo');
+
+            if ($person->save()) {
+                SiteController::addFlash('success', Yii::t('app', '{name} has been successfully edited.', ['name' => $person->fullname]));
+                return $this->redirect(['/person']);
+            } else {
+                SiteController::FlashErrors($person);
+            }
         }
 
         return $this->render('form', [
@@ -56,13 +77,36 @@ class PersonController extends BaseController {
         ]);
     }
 
-    public function actionDelete($id) {
+    public function actionDelete($id)
+    {
         $person = Person::findOne(['id' => $id]);
+
+        if (!$person || $person->coach_id != Yii::$app->user->id) {
+            throw new \yii\web\ForbiddenHttpException(Yii::t('app', 'Your not allowed to access this page.'));
+        }
+
         if ($person->delete()) {
             SiteController::addFlash('success', Yii::t('app', '{name} has been successfully deleted.', ['name' => $person->fullname]));
             return $this->redirect(['/person']);
         } else {
             SiteController::FlashErrors($person);
+        }
+    }
+
+    public static function upload($model, $attr)
+    {
+        $file = \yii\web\UploadedFile::getInstance($model, $attr);
+
+        if ($file && $model->validate()) {
+            $fileName = uniqid($attr) . '.' . $file->extension;
+            $filePath = Yii::getAlias('@webroot/photos/' . $fileName);
+
+            $file->saveAs($filePath);
+            $model->$attr = $fileName;
+
+            return true;
+        } else {
+            return false;
         }
     }
